good

+++ source/ClassFuel.cpp
@@ -8,7 +8,7 @@ ClassFuel::ClassFuel()
 
 ClassFuel::ClassFuel(Sint32 x, Sint32 y)
 {
-    this->working_mixture = { x, y, 0, 0 };
+    this->working_mixture = { static_cast<float>(x), static_cast<float>(y), 0, 0 };
     this->wm_color = { 0x00, 0x80, 0xFF };
 }
